+++ src/ui/src/js/services/utility_service.js
@@ -134,9 +134,6 @@ export default function utilityService($rootScope, $http) {
     getVersion: () => {
       return $http.get('version');
     },
-    getNamespaces: () => {
-      return $http.get('api/v1/namespaces');
-    },
     getIcon: (iconName) => {
       if (iconName === undefined || iconName == null || iconName == '') {
         if (